requirements.txt
otter-grader~=3.1.4
jupyterhub~=1.4.1
sqlalchemy~=1.4.25
stdio-proxy
setuptools~=58.1.0
tornado~=6.1
python-on-whales~=0.26.0